requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
funcsigs==1.0.2
h5py==2.7.0
itsdangerous==0.24
Jinja2==2.9.6
Keras==2.0.4
MarkupSafe==1.0
mock==2.0.0
mysql-connector==2.1.4
numpy==1.12.1
packaging==16.8
pbr==3.0.0
protobuf==3.3.0
pyparsing==2.2.0
PyYAML==3.12
scipy==0.19.0
six==1.10.0
SQLAlchemy==1.1.9
tensorflow==1.1.0
Theano==0.9.0
Werkzeug==0.12.2
